Add files via upload
The calculation of one dimensional (1D) electromagnetic (EM) responses of dipolar sources is considered as a fundamental step in a number of geophysical applications, ranging from EM sensitivity analysis, to high dimensional EM forward modeling and inversion. We develop an open source program for a complete treatment of EM responses due to various dipolar sources (i.e., vertical and horizontal electrical dipoles, and vertical and horizontal magnetic dipoles), in which the transmitter-receiver geometry is arbitrary. A complete and easy-to-follow derivation of the EM field solution for all traditional dipolar sources, at any depth, is derived based on the vector potential differential equations. The quadrature-with-extrapolation (QWE) algorithm is adapted to calculate the integration of the Bessel functions involved in EM field calculation, to produce EM responses with a desired accuracy. The accuracy of our program is verified by comparing the calculated results with the published data and analytic solution. A layered-model is designed to test the capability of our program for all traditional dipolar sources., by placing the receiver at different depths. In the end, we apply our algorithm to marine and airborne control source electromagnetic explorations. The results indicate that our program is accurate and can be used to calculate 1D EM responses for general EM problems.